packages/pytest-localostack/pyproject.toml
[project]
name = "pytest-localostack"
version = "0.1.0"
description = "Pytest plugin for LocalOStack - OpenStack API emulator fixtures"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "pytest>=8.0",
    "localostack>=0.1.0",
    "openstacksdk>=4.0.0",
    "httpx>=0.27",
]
authors = [{name = "Jaehan Byun"}]
classifiers = [
    "Framework :: Pytest",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Testing",
]
keywords = ["pytest", "openstack", "localostack", "testing"]

[project.entry-points.pytest11]
localostack = "pytest_localostack.plugin"

[project.urls]
Homepage = "https://github.com/jaehanbyun/localostack"
Repository = "https://github.com/jaehanbyun/localostack"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pytest_localostack"]
